+++ Arkanoid - SoQueNao/Barra.h
@@ -1,0 +1,23 @@
+#pragma once
+#include "Rectangulo.h"
+#include <SFML\Graphics.hpp>
+using namespace sf;
+
+
+class Barra :
+	public Rectangulo
+{
+public:
+
+
+	Barra(float mX, float mY);
+	 void update(FrameTime mFT);
+	
+private:
+	float larguraBarra = 100.f * (larguraJanela * 0, 03.f);
+	float alturaBarra = 18.f * (alturaJanela * 0, 02.f);
+	float velocidadeBarra = 1.0f * (alturaJanela * 0, 01.f);
+	Vector2f Tamanho_barra;
+	Vector2f velocidade;
+};
+
