+++ Arkanoid - SoQueNao/Config.cpp
@@ -1,0 +1,13 @@
+#include "Config.h"
+
+
+
+Config::Config()
+{
+}
+
+larguraJanela = VideoMode::getDesktopMode().width;
+alturaJanela= VideoMode::getDesktopMode().height;
+Config::~Config()
+{
+}
