+++ Arkanoid - SoQueNao/Config.h
@@ -1,0 +1,15 @@
+#pragma once
+#include <SFML/Graphics.hpp>
+using namespace sf;
+class Config
+{
+	unsigned int larguraJanela; 
+	unsigned int alturaJanela; 
+	
+
+public:
+	Config();
+	
+	~Config();
+};
+
